// estela-web/src/services/api/generated-api/models/ConnectionList.ts
/* tslint:disable */
/* eslint-disable */
/**
 * estela API v1.0 Documentation
 * estela API Swagger Specification
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ConnectionList
 */
export interface ConnectionList {
    /**
     * Connection ID.
     * @type {number}
     * @memberof ConnectionList
     */
    readonly cid?: number;
    /**
     * Connection name.
     * @type {string}
     * @memberof ConnectionList
     */
    name: string;
    /**
     * Connection type.
     * @type {string}
     * @memberof ConnectionList
     */
    connType: string;
    /**
     * Human readable connection type.
     * @type {string}
     * @memberof ConnectionList
     */
    readonly connTypeDisplay?: string;
    /**
     * Creation date.
     * @type {Date}
     * @memberof ConnectionList
     */
    readonly created?: Date;
}

export function ConnectionListFromJSON(json: any): ConnectionList {
    return ConnectionListFromJSONTyped(json, false);
}

export function ConnectionListFromJSONTyped(json: any, ignoreDiscriminator: boolean): ConnectionList {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'cid': !exists(json, 'cid') ? undefined : json['cid'],
        'name': json['name'],
        'connType': json['conn_type'],
        'connTypeDisplay': !exists(json, 'conn_type_display') ? undefined : json['conn_type_display'],
        'created': !exists(json, 'created') ? undefined : (new Date(json['created'])),
    };
}

export function ConnectionListToJSON(value?: ConnectionList | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'name': value.name,
        'conn_type': value.connType,
    };
}